requirements.txt
azure-functions~=1.7.2